add checks to avoid errors when uninstalling apps with custom repository

+++ apps-development/beekeeper-studio/uninstall.sh
@@ -1,5 +1,9 @@
-sudo apt purge -y --autoremove beekeeper-studio
+# uninstall only if it's already installed
+if apt search beekeeper-studio | grep installed &> /dev/null
+then
+  sudo apt purge -y --autoremove beekeeper-studio
+fi
 
-sudo rm /etc/apt/sources.list.d/beekeeper-studio-app.list
-
-sudo rm /usr/share/keyrings/beekeeper.gpg
+# remove repository
+sudo rm -f /etc/apt/sources.list.d/beekeeper-studio-app.list
+sudo rm -f /usr/share/keyrings/beekeeper.gpg
