+++ apps-development/bruno/uninstall.sh
@@ -1,5 +1,9 @@
-sudo apt purge -y --autoremove bruno
+# uninstall only if it's already installed
+if apt search bruno | grep installed &> /dev/null
+then
+  sudo apt purge -y --autoremove bruno
+fi
 
-sudo rm /etc/apt/sources.list.d/bruno.list
-
-sudo rm /etc/apt/keyrings/bruno.gpg
+# remove repository
+sudo rm -f /etc/apt/sources.list.d/bruno.list
+sudo rm -f /etc/apt/keyrings/bruno.gpg
