+++ apps-development/docker/uninstall.sh
@@ -1,12 +1,17 @@
+# remove user from docker group
 sudo usermod -rG docker $USER
 
-sudo apt purge -y --autoremove \
-  docker-ce \
-  docker-ce-cli \
-  containerd.io \
-  docker-buildx-plugin \
-  docker-compose-plugin
+# ininstall if it's already installed
+if apt search docker-ce | grep installed &> /dev/null
+then
+  sudo apt purge -y --autoremove \
+    docker-ce \
+    docker-ce-cli \
+    containerd.io \
+    docker-buildx-plugin \
+    docker-compose-plugin
+fi
 
-sudo rm /etc/apt/sources.list.d/docker.list
-
-sudo rm /etc/apt/keyrings/docker.asc
+# remove repository
+sudo rm -f /etc/apt/sources.list.d/docker.list
+sudo rm -f /etc/apt/keyrings/docker.asc
